게시물 삭제시 권한체크

+++ src/main/java/com/sbs/exam/demo/controller/UsrArticleController.java
@@ -31,7 +31,7 @@ public class UsrArticleController {
 		}
 		
 		if (isLogined == false) {
-			return ResultData.from("F-1", "로그인을 해주세요!");
+			return ResultData.from("F-A", "로그인을 해주세요!");
 		}
 
 		if (Ut.empty(title)) {
@@ -68,13 +68,27 @@ public class UsrArticleController {
 
 	@RequestMapping("usr/article/doDelete")
 	@ResponseBody
-	public ResultData<Integer> doDelete(int id) {
+	public ResultData<Integer> doDelete(HttpSession httpSession, int id) {
+		boolean isLogined = false;
+		int isLoginedMemberId = 0;
+		
+		if (httpSession.getAttribute("loginedMemberId") != null) {
+			isLogined = true;
+			isLoginedMemberId = (int) httpSession.getAttribute("loginedMemberId");
+		}
+		
+		if (httpSession.getAttribute("loginedMemberId") == null) {
+			return ResultData.from("F-A", "로그인을 해주세요!");
+		}
 		Article article = articleService.getArticle(id);
-
-		if (article == null) {
-			return ResultData.from("F-1", Ut.f("%d번 게시물은 존재하지 않습니다.", id));
+		if(article  == null) {
+			return ResultData.from("F-2", Ut.f("%d번 게시물이 존재하지 않습니다.", id));
+		}
+		if (article.getMemberId() != isLoginedMemberId) {
+			return ResultData.from("F-2", "해당 게시물에 대한 권한이 없습니다.");
 		}
 		articleService.deleteArticle(id);
+
 		return ResultData.from("S-1", Ut.f("%d번 게시물을 삭제했습니다.", id), id);
 	}
